refactor: move matchDeprecated function to utils and update its implementation
