+++ .dumi/theme/common/ComponentChangelog/ComponentChangelog.tsx
@@ -3,18 +3,12 @@ import { BugOutlined } from '@ant-design/icons';
 import { Button, Drawer, Flex, Grid, Popover, Tag, Timeline, Typography } from 'antd';
 import type { TimelineItemProps } from 'antd';
 import { createStyles } from 'antd-style';
-import semver from 'semver';
 
-import deprecatedVersions from '../../../../BUG_VERSIONS.json';
 import useFetch from '../../../hooks/useFetch';
 import useLocale from '../../../hooks/useLocale';
 import useLocation from '../../../hooks/useLocation';
+import { matchDeprecated } from '../../utils';
 import Link from '../Link';
-
-interface MatchDeprecatedResult {
-  match?: string;
-  reason: string[];
-}
 
 interface ChangelogInfo {
   version: string;
@@ -22,17 +16,6 @@ interface ChangelogInfo {
   refs: string[];
   contributors: string[];
   releaseDate: string;
-}
-
-function matchDeprecated(v: string): MatchDeprecatedResult {
-  const match = Object.keys(deprecatedVersions).find((depreciated) =>
-    semver.satisfies(v, depreciated),
-  );
-  const reason = deprecatedVersions[match as keyof typeof deprecatedVersions] || [];
-  return {
-    match,
-    reason: Array.isArray(reason) ? reason : [reason],
-  };
 }
 
 const useStyle = createStyles(({ token, css }) => ({
