+++ .dumi/theme/utils.ts
@@ -1,5 +1,7 @@
 import flatten from 'lodash/flatten';
 import flattenDeep from 'lodash/flattenDeep';
+import deprecatedVersions from '../../BUG_VERSIONS.json';
+import semver from 'semver';
 
 import themeConfig from './themeConfig';
 
@@ -22,6 +24,11 @@ interface ModuleDataItem {
 
 interface Orders {
   [key: string]: number;
+}
+
+interface MatchDeprecatedResult {
+  match?: string;
+  reason: string[];
 }
 
 export function getMenuItems(
@@ -199,6 +206,17 @@ export function getMetaDescription(jml?: any[] | null) {
       }),
   ).find((p) => p && typeof p === 'string' && !COMMON_TAGS.includes(p)) as string;
   return paragraph;
+}
+
+export function matchDeprecated(v: string): MatchDeprecatedResult {
+  const match = Object.keys(deprecatedVersions).find((depreciated) =>
+    semver.satisfies(v, depreciated),
+  );
+  const reason = deprecatedVersions[match as keyof typeof deprecatedVersions] || [];
+  return {
+    match,
+    reason: Array.isArray(reason) ? reason : [reason],
+  };
 }
 
 export const getThemeConfig = () => themeConfig;
